requirements.txt
aiogram
aiosqlite
supabase
python-dotenv
pydantic
pydantic-settings
